Cleanup in MoreViewActions and MoreViewObservables
